Secure CollectionJob resolvers

+++ server/resolvers/deleteCollectionJobResolver.js
@@ -1,8 +1,14 @@
+const { AuthenticationError } = require('apollo-server-core');
 const {
     deleteCollectionJob
 } = require('../models/services/CollectionJobService');
 
-const deleteCollectionJobResolver = async (_, { id }) =>
-    await deleteCollectionJob(id);
+const deleteCollectionJobResolver = async (_, { id }, context) => {
+    const { user } = context;
+    if (!user?.roles.find(role => role.name === 'ADMIN')) {
+        throw new AuthenticationError();
+    }
+    return deleteCollectionJob(id);
+};
 
 module.exports = deleteCollectionJobResolver;
